CHORE: 로직들 추가
