+++ constants/queryKeys.ts
@@ -1,3 +1,4 @@
 export const ABANDONMENT_QUERY_KEY = 'abandonment';
 export const ABANDONMENTS_QUERY_KEY = 'abandonments';
 export const BANNERS_QUERY_KEY = 'banners';
+export const SHELTER_QUERY_KEY = 'shelter';
