+++ mocks/handlers.ts
@@ -1,3 +1,5 @@
-import { abandonment, abandonments, banners } from './service';
+import { abandonment, abandonments } from './abandonments';
+import { banners } from './banners';
+import { shelter } from './shelters';
 
-export const handlers = [banners, abandonments, abandonment];
+export const handlers = [banners, abandonments, abandonment, shelter];
